Long to integer conversion

+++ book-network-api/src/main/java/com/aed/booknetwork/config/BeansConfig.java
@@ -32,7 +32,7 @@ public class BeansConfig {
     }
 
     @Bean
-    public AuditorAware<Long> auditorAware() {
+    public AuditorAware<Integer> auditorAware() {
         return new ApplicationAuditAware();
     }
